Remove monkeypatched logger
Rather than trap all the log messages in a fake data structure, it’s
probably easier and more flexible to customise the log appender in the
`.smithrc` for test runs.

Setting the appender to `StringIo` will use a buffer object to capture
log messages, and they can be dumped (or asserted against) within a unit
test block in RSpec. In other words, everything that FakeSmith currently provides can
be achieved directly within the `Logging` gem itself.